Code cleanup, will probably pull this into the master branch soon

+++ DenaryBinaryHexOctalGUIVer/Converter.cs
@@ -18,13 +18,15 @@ namespace DenaryBinaryHexOctalGUIVer
             int raw = 0;
             StringBuilder output = new StringBuilder();
 
+            //calculate the raw value
             for (int i = 0; i < input.Length; i++)
             {
                 char letter = input[input.Length - i - 1];
-                raw += ((Convert.ToInt32(letter) - (Char.IsDigit(letter) ? 48 : 55)) * (int)Math.Pow((int)from, i));
+                raw += (Convert.ToInt32(letter) - (Char.IsDigit(letter) ? 48 : 55)) * (int)Math.Pow((int)from, i);
             }
 
             bool passedBeginningZeros = false;
+
             for (int i = 32; i >= 0; i--)
             {
                 //How many times the current raw number divides by
@@ -33,7 +35,7 @@ namespace DenaryBinaryHexOctalGUIVer
                 //Stop it turning into 0000000000000000001010 or something similar
                 if (divided > 0) passedBeginningZeros = true;
 
-                if (divided > 9) output.Append((char) ((int) divided + 55)); //append appropriate letter for > base 10 using the fact that A (10) is ASCII code 55 + 10 = 65 TODO: Figure out what to do for > base 36
+                if (divided > 9) output.Append((char) (divided + 55)); //append appropriate letter for > base 10 using the fact that A (10) is ASCII code 55 + 10 = 65 TODO: Figure out what to do for > base 36
                 else if (passedBeginningZeros) output.Append(divided);
 
                 //reduce the raw value
@@ -54,6 +56,7 @@ namespace DenaryBinaryHexOctalGUIVer
             for (int i = 0; i < input.Length; i++)
             {
                 char letter = input.ToUpper()[i];
+
                 if (Char.IsLetterOrDigit(letter))
                 {
                     if (Char.IsDigit(letter) && Convert.ToInt32(letter) - 48 > ((int) numBase) - 1) return false;
